refactor(tests): split test_validate.py and planner_test.go into focused modules
Removes both files from GRANDFATHERED_FILES so the file size policy is
enforced going forward.

planner_test.go (904 lines) → 6 focused files:
- planner_helpers_test.go  — newTestPlanner, writeTempYAML helpers
- planner_parse_test.go    — Parse/* validation tests
- planner_dag_test.go      — ValidateDAG cycle/dep tests
- planner_plan_test.go     — Plan/* + Pipeline integration test
- planner_regex_test.go    — stepIDRegex / ResourceIDRegex tests
- planner_step_limits_test.go — Parse/StepLimits/* tests

test_validate.py (1064 lines) → 3 focused files:
- test_validate_agent_schema.py — agent types, IDs, safety net, multi-agent
- test_validate_persona.py      — behavior, autonomy, memory, range constraints
- test_validate_workflow.py     — workflows, channels, edge cases, structured errors
