+++ FILEMAP.md
@@ -1,16 +1,16 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-04-25  
-> **426** tracked files
+> **433** tracked files
 
 ## Summary
 
 | Top-level directory | Files |
 |---------------------|-------|
-| `internal/` | 113 |
+| `internal/` | 118 |
 | `docs/` | 81 |
 | `agents/` | 77 |
-| `tests/` | 65 |
+| `tests/` | 67 |
 | `(root)` | 20 |
 | `scripts/` | 17 |
 | `.github/` | 15 |
@@ -26,8 +26,8 @@
 
 | Extension | Count |
 |-----------|-------|
-| `.py` | 153 |
-| `.go` | 110 |
+| `.py` | 155 |
+| `.go` | 115 |
 | `.md` | 97 |
 | `.yaml` | 25 |
 | `.rs` | 9 |
@@ -381,7 +381,12 @@
 │   │       └── stderr.go
 │   ├── planner/
 │   │   ├── planner.go
-│   │   ├── planner_test.go
+│   │   ├── planner_dag_test.go
+│   │   ├── planner_helpers_test.go
+│   │   ├── planner_parse_test.go
+│   │   ├── planner_plan_test.go
+│   │   ├── planner_regex_test.go
+│   │   ├── planner_step_limits_test.go
 │   │   ├── resolve_test.go
 │   │   └── testdata/
 │   │       ├── cycle_complex.yaml
@@ -540,7 +545,9 @@
 │           ├── test_task_agent.py
 │           ├── test_tick_scheduler.py
 │           ├── test_tools.py
-│           ├── test_validate.py
+│           ├── test_validate_agent_schema.py
+│           ├── test_validate_persona.py
+│           ├── test_validate_workflow.py
 │           └── test_working_memory.py
 └── workflows/
     ├── budget-test.yaml
