+++ scripts/checks/file_size.py
@@ -90,14 +90,12 @@ GRANDFATHERED_FILES: frozenset[str] = frozenset({
     # Go test files that exceeded the 500-line limit when the **/*_test.go
     # blanket exclusion was removed. These are tracked for targeted follow-up
     # splits. Do not add new entries — split instead.
-    "internal/planner/planner_test.go",
     "internal/state/state_test.go",
     "internal/cost/cost_test.go",
     "internal/observability/zapenc/encoder_test.go",
     # Python test files that exceeded the 500-line limit when the tests/**
     # blanket exclusion was removed (v0.2 baseline). These are tracked for
     # targeted follow-up splits. Do not add new entries — split instead.
-    "tests/unit/python/test_validate.py",
     "tests/unit/python/test_server.py",
     "tests/unit/python/test_chat_servicer.py",
     "tests/unit/python/test_memory_tools.py",
